[129395] Moved codgen TypeResolver to java.core so that it can be used anywhere, and added it to editdomain so that anyone can get the current typeresolver from the edit domain and do quick resolutions.
This was needed because property editors where returning init strings of static field accesses, and without the type resolver there was no way to know what was the class part and what was the field part. This would screw load from cache up because it would try to resolve "java.lang.Color.blue" as a class and it couldn't do it. By using the type resolver it is smart enough now to know this is a field access of blue from class of java.awt.Color.